requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-chess==1.999
requests==2.31.0